w3c-reports/12092014/mac/runned-tests/websockets/Create-Secure-url-with-space-expected.txt

CONSOLE MESSAGE: Invalid url for WebSocket ws://web platform.test:49001/echo

PASS W3C WebSocket API - Create Secure WebSocket - Pass a URL with a space - SYNTAX_ERR should be thrown 
